+++ webviz_subsurface/ert_jobs/jobs.py
@@ -36,7 +36,8 @@ def installable_jobs() -> dict:
         resource_filename("webviz_subsurface", "ert_jobs/config_jobs")
     )
     return {
-        "WELL_CONNECTION_STATUS": str(resource_directory / "WELL_CONNECTION_STATUS")
+        "WELL_CONNECTION_STATUS": str(resource_directory / "WELL_CONNECTION_STATUS"),
+        "SMRY2ARROW": str(resource_directory / "SMRY2ARROW"),
     }